"check_style_score -> 0 of 1 passed. style score: 0/1 "

+++ cspp2 assignments/M4/Assignment-3/Solution.java
@@ -18,7 +18,7 @@ public class Solution
 		}
 	}
 		public static int binaryToDecimal(String a) {
-			double d = Integer.parseInt(a);
+			int d = Integer.parseInt(a);
 			double r = 0;
 			double digit = 0;
 			int c = a.length();
